rename to gnmi
