+++ pkg/southbound/gnmi/client.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v070
+package gnmi
 
 import (
 	"context"
